pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "salute_speech"
version = "1.3.0"
description = "Sber Salute Speech API"
readme = "README.md"
authors = [
    {name = "Maxim Moroz", email = "maxim.moroz@gmail.com"},
]
license = {file = "LICENSE"}
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
]
keywords = ["speech"]
requires-python = ">=3.8"
dependencies = [
    "Click>=8.1.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "pydub>=0.25.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.7.0",
    "numpy>=1.20.0",
]

[project.urls]
Homepage = "https://github.com/mmua/salute_speech"
Repository = "https://github.com/mmua/salute_speech.git"

[project.scripts]
salute_speech = "salute_speech:cli"

[tool.setuptools]
packages = ["salute_speech"]
include-package-data = true
zip-safe = false

[tool.setuptools.package-data]
salute_speech = ["conf/*"]
